src/backend/requirements.txt
# FastAPI 核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 資料庫
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0  # PostgreSQL async driver
aiosqlite==0.19.0  # SQLite async driver

# 認證與安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# AI/ML
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78
numpy==1.24.4
Pillow==10.1.0
scikit-learn==1.3.2

# 資料處理
pandas==2.1.4
pydantic==2.5.1
pydantic-settings==2.1.0

# HTTP 客戶端
httpx==0.25.2
requests==2.31.0

# 快取
redis==5.0.1
aioredis==2.0.1

# 開發工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# 監控與日誌
prometheus-client==0.19.0
structlog==23.2.0

# 其他工具
python-dotenv==1.0.0
Jinja2==3.1.2